Separated process implementation into separate file.
