+++ Sources/VHDLGenerator/stateGenerator/AsynchronousBlock+stateGenerator.swift
@@ -200,40 +200,3 @@ extension AsynchronousBlock {
     }
 
 }
-
-extension ProcessBlock {
-
-    init?<T>(
-        stateGeneratorFor state: State, in representation: T, maxExecutionSize: Int? = nil
-    ) where T: MachineVHDLRepresentable {
-        guard let checkForDuplicates = WhenCase(
-            stateGeneratorCheckForDuplicatesFor: state, in: representation, maxExecutionSize: maxExecutionSize
-        ) else {
-            return nil
-        }
-        let clk = representation.machine.clocks[representation.machine.drivingClock].name
-        self.init(
-            sensitivityList: [clk],
-            code: .ifStatement(block: .ifStatement(
-                condition: .conditional(condition: .edge(value: .rising(expression: .reference(
-                    variable: .variable(reference: .variable(name: clk))
-                )))),
-                ifBlock: .caseStatement(block: CaseStatement(
-                    condition: .reference(variable: .variable(reference: .variable(name: .internalState))),
-                    cases: [
-                        .stateGeneratorInitial,
-                        .stateGeneratorCheckForJob,
-                        .stateGeneratorWaitForRunnerToStart,
-                        .stateGeneratorWaitForRunnerToFinish,
-                        .stateGeneratorWaitForCacheToStart,
-                        checkForDuplicates,
-                        WhenCase(stateGeneratorAddToStatesFor: state, in: representation),
-                        .stateGeneratorWaitForCacheToEnd,
-                        .othersNull
-                    ]
-                ))
-            ))
-        )
-    }
-
-}
